# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Builders MVP API",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/v1/admin/therapist-verifications/{verificationId}/reject": {
            "post": {
                "tags": [
                    "admin-therapist-verification-controller"
                ],
                "operationId": "reject",
                "parameters": [
                    {
                        "name": "verificationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RejectTherapistVerificationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTherapistVerificationResponse"
                                }
                            }
                        }
                    }
                },
                "description": "[관리자] 치료사 인증 반려 (사유 포함)",
                "summary": "[관리자] 치료사 인증 반려 (사유 포함)"
            }
        },
        "/api/v1/admin/therapist-verifications/{verificationId}/approve": {
            "post": {
                "tags": [
                    "admin-therapist-verification-controller"
                ],
                "operationId": "approve",
                "parameters": [
                    {
                        "name": "verificationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTherapistVerificationResponse"
                                }
                            }
                        }
                    }
                },
                "description": "[관리자] 치료사 인증 승인 (USER → THERAPIST 역할 승격)",
                "summary": "[관리자] 치료사 인증 승인 (USER → THERAPIST 역할 승격)"
            }
        },
...
        "/api/v1/admin/therapist-verifications": {
            "get": {
                "tags": [
                    "admin-therapist-verification-controller"
                ],
                "operationId": "getVerifications",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTherapistVerificationPageResponse"
                                }
                            }
                        }
                    }
                },
                "description": "[관리자] 치료사 인증 목록 조회 (상태별 필터, 페이지네이션)",
                "summary": "[관리자] 치료사 인증 목록 조회 (상태별 필터, 페이지네이션)"
            }
        },
        "/api/v1/admin/therapist-verifications/{verificationId}/image": {
            "get": {
                "tags": [
                    "admin-therapist-verification-controller"
                ],
                "operationId": "downloadVerificationImage",
                "parameters": [
                    {
                        "name": "verificationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                },
                "description": "[관리자] 치료사 인증 이미지 다운로드",
                "summary": "[관리자] 치료사 인증 이미지 다운로드"
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiResponseTherapistVerificationResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/TherapistVerificationResponse"
                    }
                }
            },
            "TherapistVerificationResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userEmail": {
                        "type": "string"
                    },
                    "userNickname": {
                        "type": "string"
                    },
                    "licenseCode": {
                        "type": "string"
                    },
                    "licenseImageOriginName": {
                        "type": "string"
                    },
                    "licenseImageDownloadUrl": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "APPROVED",
                            "REJECTED"
                        ]
                    },
                    "reviewedById": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "reviewedByNickname": {
                        "type": "string"
                    },
                    "reviewedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "rejectReason": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "RejectTherapistVerificationRequest": {
                "type": "object",
                "properties": {
                    "rejectReason": {
                        "type": "string"
                    }
                },
                "required": [
                    "rejectReason"
                ]
            },
...
            "ApiResponseTherapistVerificationPageResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/TherapistVerificationPageResponse"
                    }
                }
            },
            "TherapistVerificationPageResponse": {
                "type": "object",
                "properties": {
                    "verifications": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TherapistVerificationResponse"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "hasNext": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}